rebase / spread components to folder
