+++ src/renderer/modules/webpack/common/fluxDispatcher.ts
@@ -1,6 +1,14 @@
+import { ModuleExports } from "../../../../types/discord";
 import { filters, waitForModule } from "..";
 
-// @todo: type this
-export default await waitForModule(
+type FluxCallback = (event?: { [ index: string ]: unknown }) => void;
+export type FluxDispatcher = ModuleExports & {
+  _subscriptions: { [ index: string ]: Set<FluxCallback>};
+  dispatch: (event: { type: string, [ index: string ]: unknown }) => void;
+  subscribe: (eventKey: string, callback: FluxCallback) => void;
+  unsubscribe: (eventKey: string, callback: FluxCallback) => void;
+};
+
+export default await waitForModule<FluxDispatcher>(
   filters.byProps("_currentDispatchActionType", "_processingWaitQueue"),
 );
