+++ src/renderer/modules/webpack/common/index.ts
@@ -29,8 +29,8 @@ importTimeout("channels", import("./channels"), (mod) => (channels = mod.default
 export let constants: typeof import("./constants");
 importTimeout("constants", import("./constants"), (mod) => (constants = mod));
 
-// export let contextMenu: typeof import("./contextMenu");
-// importTimeout("contextMenu", import("./contextMenu"), (mod) => (contextMenu = mod.default));
+export let contextMenu: typeof import("./contextMenu").default;
+importTimeout("contextMenu", import("./contextMenu"), (mod) => (contextMenu = mod.default));
 
 export let flux: typeof import("./flux").default;
 importTimeout("flux", import("./flux"), (mod) => (flux = mod.default));
@@ -51,8 +51,8 @@ importTimeout("hljs", import("./hljs"), (mod) => (hljs = mod.default));
 export let messages: typeof import("./messages").default;
 importTimeout("messages", import("./messages"), (mod) => (messages = mod.default));
 
-// export let modal: typeof import("./modal").default;
-// importTimeout("modal", import("./modal"), (mod) => (modal = mod.default));
+export let modal: typeof import("./modal").default;
+importTimeout("modal", import("./modal"), (mod) => (modal = mod.default));
 
 export let React: typeof import("./react").default;
 importTimeout("React", import("./react"), (mod) => (React = mod.default));
